update/refactor tests to use test attrs
