+++ gatsby/cypress/e2e/mobile.cy.js
@@ -9,27 +9,6 @@ mobile.devices.forEach((device) => {
         cy.wait(waitTime)
       })
 
-      it('page', () => {
-        const sections = cy.get('section')
-        sections.should('have.length', 6)
-
-        // get 2nd section
-        const div = sections.eq(1).get('div')
-        const img_top = div.eq(0).get('img').eq(0)
-        const img_bottom = div.eq(0).get('img').eq(1)
-
-        // without this sometimes they don't show up in the screenshot
-        img_top.should('be.visible')
-        img_bottom.should('be.visible')
-
-        // wait for rendering to finish
-        cy.wait(waitTime)
-
-        cy.compareSnapshot('page_mobile_' + device + '_' + orientation, {
-          errorThreshold: errorThreshold
-        })
-      })
-
       it('navbar non-expanded', () => {
         const header = cy.get('header')
         const nav = cy.get('header nav')
@@ -40,6 +19,12 @@ mobile.devices.forEach((device) => {
         } else {
           svg.should('be.visible')
         }
+      })
+
+      it('page screenshot', () => {
+        cy.compareSnapshot('page_mobile_' + device + '_' + orientation, {
+          errorThreshold: errorThreshold
+        })
       })
     })
   })
